Workload state refactored to a separate class
